+++ src/AICentral/ConsumerAuth/AICentralJWT/AICentralJwtProviderResponse.cs
@@ -1,0 +1,6 @@
+﻿namespace AICentral.ConsumerAuth.AICentralJWT;
+
+public class AICentralJwtProviderResponse
+{
+    public AICentralApiKeyToken[] Tokens { get; set; } = default!;
+}