+++ src/AICentralTests/InBuiltJwtAuth/the_inbuild_auth_provider.cs
@@ -27,13 +27,16 @@ public class the_inbuild_auth_provider : IClassFixture<TestWebApplicationFactory
     public async Task can_create_and_use_jwt_tokens()
     {
         var request = new HttpRequestMessage(HttpMethod.Post,
-            "https://not-used-for-tokens/aicentraljwt/azure-openai-to-azure-with_custom_jwt/tokens")
+            "https://not-used-for-tokens/aicentraljwt/jwt1/tokens")
         {
-            Content = JsonContent.Create(new
+            Content = JsonContent.Create(new TokenRequest()
             {
                 Names = new[] { "Hacker1", "Hacker2", "Hacker3", "Hacker4" },
-                ValidPipelines = new[] { "azure-openai-to-azure-with_custom_jwt.localtest.me-pipeline" },
-                ValidFor = "00:08:00"
+                ValidPipelines = new Dictionary<string, string[]>
+                {
+                    ["azure-openai-to-azure-with_custom_jwt.localtest.me-pipeline"] = ["random"],
+                },
+                ValidFor = TimeSpan.FromHours(8)
             })
         };
         request.Headers.Add("api-key", "fake-admin-key");
@@ -70,6 +73,69 @@ public class the_inbuild_auth_provider : IClassFixture<TestWebApplicationFactory
 
         var response = await client.GetChatCompletionsAsync(new ChatCompletionsOptions("random",
                 [new ChatRequestSystemMessage("Test")]));
+
+        response.GetRawResponse().Status.ShouldBe(200);
+
+    }
+
+    [Fact]
+    public async Task refuses_to_generate_tokens_not_in_config()
+    {
+        var request = new HttpRequestMessage(HttpMethod.Post,
+            "https://not-used-for-tokens/aicentraljwt/jwt1/tokens")
+        {
+            Content = JsonContent.Create(new TokenRequest()
+            {
+                Names = new[] { "Hacker1", "Hacker2", "Hacker3", "Hacker4" },
+                ValidPipelines = new Dictionary<string, string[]>
+                {
+                    ["azure-openai-to-azure-with_custom_jwt.localtest.me-pipeline"] = ["does-not-exist"],
+                },
+                ValidFor = TimeSpan.FromHours(8)
+            })
+        };
+        request.Headers.Add("api-key", "fake-admin-key");
+
+        var tokensResponse = await _httpClient.SendAsync(request);
+        tokensResponse.StatusCode.ShouldBe(HttpStatusCode.BadRequest);
+    }
+    
+    [Fact]
+    public async Task will_generate_tokens_based_on_wildcard()
+    {
+        _factory.SeedChatCompletions(AICentralFakeResponses.Endpoint200, "random",
+            () => Task.FromResult(AICentralFakeResponses.FakeChatCompletionsResponse()));
+
+        var request = new HttpRequestMessage(HttpMethod.Post,
+            "https://not-used-for-tokens/aicentraljwt/jwt2/tokens")
+        {
+            Content = JsonContent.Create(new TokenRequest()
+            {
+                Names = ["Hacker1"],
+                ValidPipelines = new Dictionary<string, string[]>
+                {
+                    ["azure-openai-to-azure-with_custom_jwt-wildcard.localtest.me-pipeline"] = ["*"],
+                },
+                ValidFor = TimeSpan.FromHours(8)
+            })
+        };
+        request.Headers.Add("api-key", "fake-admin-key");
+
+        var tokensResponse = await _httpClient.SendAsync(request);
+        tokensResponse.StatusCode.ShouldBe(HttpStatusCode.OK);
+
+        var tokens = (await tokensResponse.Content.ReadFromJsonAsync<AICentralJwtProviderResponse>())!;
+
+        var client = new OpenAIClient(
+            new Uri("http://azure-openai-to-azure-with_custom_jwt-wildcard.localtest.me"),
+            new AzureKeyCredential(tokens.Tokens[0].ApiKeyToken),
+            new OpenAIClientOptions()
+            {
+                Transport = new HttpClientTransport(_httpClient)
+            });
+
+        var response = await client.GetChatCompletionsAsync(new ChatCompletionsOptions("random",
+            [new ChatRequestSystemMessage("Test")]));
 
         response.GetRawResponse().Status.ShouldBe(200);